requirements.txt
docx2txt
docx
streamlit
gtts
pyttsx3
pillow
translate
googletrans==4.0.0-rc1
python-docx
bs4
PyMuPDF
pytesseract
fitz
easyocr
PyPDF2
langdetect
SpeechRecognition
pydub
transformers==4.28.0
nltk
spacy
reportlab
fpdf
